requirements.txt
pymupdf==1.22.3
aioredis==2.0.1
aiohttp==3.8.4
loguru==0.7.0
peewee==3.16.2
tiktoken==0.4.0
aiofiles==23.1.0
httpx==0.24.1
pymysql2==1.3.3
python-dotenv==1.0.0
numpy==1.24.3
pydantic==1.10.9
fastapi==0.98.0
langchain==0.0.214
uvicorn==0.22.0
openai==0.27.8
pymilvus==2.2.8
